# Optimizations to updateParametersInContext()