This is a further attempt at improving #1661.  @jchodera can you try it and see if it helps?