@peastman 2018-03-20
Actually, hold on.  There's one more minor improvement I want to make.