@bas-rustenburg 2018-03-21
Going to see how it affects my benchmark results. Thanks for putting this together.