@jchodera 2018-03-23
@bas-rustenburg posted his performance data here: https://github.com/pandegroup/openmm/issues/1661#issuecomment-375405715