@peastman 2018-03-27
@bas-rustenburg I couldn't quite tell from your comments whether this change improved performance or not.  For reference, here's what I'm getting before and after this change, running on a 4 core CPU and a Titan X.

Before:

```
VerletIntegrator.step(nsteps)                                   :    0.015 s for      500 steps (   1.000 ps) :    0.030 ms / step :   1.000 x
VerletIntegrator                                                :    0.494 s for      500 steps (   1.000 ps) :    0.987 ms / step :   1.000 x
VerletIntegrator with updateParametersInContext                 :    1.688 s for      500 steps (   1.000 ps) :    3.376 ms / step :   3.421 x
LangevinIntegrator                                              :    0.482 s for      500 steps (   1.000 ps) :    0.965 ms / step :   0.977 x
LangevinIntegrator with updateParametersInContext               :    1.676 s for      500 steps (   1.000 ps) :    3.352 ms / step :   3.396 x
GHMCIntegratorIntegrator                                        :    0.000 s for      500 steps (   1.000 ps) :    0.001 ms / step :   0.001 x
GHMCIntegrator with updateParametersInContext                   :    3.791 s for      500 steps (   1.000 ps) :    7.581 ms / step :   7.681 x
```

After:

```
VerletIntegrator.step(nsteps)                                   :    0.015 s for      500 steps (   1.000 ps) :    0.030 ms / step :   1.000 x
VerletIntegrator                                                :    0.578 s for      500 steps (   1.000 ps) :    1.156 ms / step :   1.000 x
VerletIntegrator with updateParametersInContext                 :    1.435 s for      500 steps (   1.000 ps) :    2.869 ms / step :   2.482 x
LangevinIntegrator                                              :    0.484 s for      500 steps (   1.000 ps) :    0.968 ms / step :   0.838 x
LangevinIntegrator with updateParametersInContext               :    1.279 s for      500 steps (   1.000 ps) :    2.558 ms / step :   2.213 x
GHMCIntegratorIntegrator                                        :    0.000 s for      500 steps (   1.000 ps) :    0.001 ms / step :   0.001 x
GHMCIntegrator with updateParametersInContext                   :    3.322 s for      500 steps (   1.000 ps) :    6.645 ms / step :   5.748 x
```

So I'm seeing a small but still useful speedup.