@bas-rustenburg 2018-03-27
Thanks Peter.

I ran the other benchmark script. I think I can conclude it is slightly slower on our system. This is on a GeForce GTX 1080 Ti and 1 cpu core.

Before
```
(benchmark-before) rustenburg@lt01:~/openmm-protons-benchmarks/MinimalTest2:
 Tue Mar 27 16:23:02 [0] $ python minimal_test.py
openmm 6b32fc32dddada6418f52da2283986a91f69f8d4
System has 50442 particles
VerletIntegrator.step(nsteps)                                   :    2.877 s for      500 steps (   1.000 ps) :    5.753 ms / step :   1.000 x
VerletIntegrator                                                :    2.893 s for      500 steps (   1.000 ps) :    5.785 ms / step :   1.000 x
VerletIntegrator with updateParametersInContext                 :   20.454 s for      500 steps (   1.000 ps) :   40.908 ms / step :   7.071 x
LangevinIntegrator                                              :    2.913 s for      500 steps (   1.000 ps) :    5.826 ms / step :   1.007 x
LangevinIntegrator with updateParametersInContext               :   20.109 s for      500 steps (   1.000 ps) :   40.218 ms / step :   6.952 x
GHMCIntegratorIntegrator                                        :    5.410 s for      500 steps (   1.000 ps) :   10.820 ms / step :   1.870 x
GHMCIntegrator with updateParametersInContext                   :   26.756 s for      500 steps (   1.000 ps) :   53.511 ms / step :   9.250 x
```
After
```(benchmark-after) rustenburg@lt01:~/openmm-protons-benchmarks/MinimalTest2:
 Tue Mar 27 16:26:24 [0] $ python minimal_test.py
openmm af57bf9852d7d7de958a2c0db75213059a79e35f
System has 50442 particles
VerletIntegrator.step(nsteps)                                   :    2.863 s for      500 steps (   1.000 ps) :    5.726 ms / step :   1.000 x
VerletIntegrator                                                :    2.902 s for      500 steps (   1.000 ps) :    5.804 ms / step :   1.000 x
VerletIntegrator with updateParametersInContext                 :   21.098 s for      500 steps (   1.000 ps) :   42.195 ms / step :   7.270 x
LangevinIntegrator                                              :    2.910 s for      500 steps (   1.000 ps) :    5.821 ms / step :   1.003 x
LangevinIntegrator with updateParametersInContext               :   21.009 s for      500 steps (   1.000 ps) :   42.019 ms / step :   7.240 x
GHMCIntegratorIntegrator                                        :    5.283 s for      500 steps (   1.000 ps) :   10.567 ms / step :   1.821 x
GHMCIntegrator with updateParametersInContext                   :   27.696 s for      500 steps (   1.000 ps) :   55.391 ms / step :   9.544 x
```