@peastman 2018-03-27
Right, it will be a little slower with only one core.  Part of the optimization comes from parallelizing it.  Is there a reason you're only using one core?