@jchodera 2018-03-27
Ah! We usually set `OPENMM_CPU_THREADS` to `1` to run on our batch queue system so we can request an equal number of thread-slots as GPUs from the batch queue system. We can look into how to request multiple thread-slots per GPU (on the same node) to let us use multiple cores, but we hadn't done this before because we didn't know there was an advantage in doing so!