@bas-rustenburg 2018-03-27
As John notes, we can only specify GPU requests in a `gpus per cpu` way on our cluster because of poor GPU support in the LSF queue software. We, and users from other groups have requested multiple times that they implement this, but so far no luck with the sysadmins. Until then, I usually am stuck running on 1 cpu and 1 gpu.