@peastman 2018-03-27
There aren't a lot of parallel operations in the CUDA platform.  I think the only other one might be that `getState()` is a little faster, since it parallelizes the work of translating all the positions into the proper periodic box.

Shall I go ahead and merge this?  I expect multiple cores to be the more common case in practice, and the speedup when you have multiple cores seems to be larger than the slowdown when you don't.