@jchodera 2018-03-27
Seems fine to merge!